SKALE-1450 Fix name conflicts, particularly with u256
