+++ node/ConsensusEngine.h
@@ -27,6 +27,7 @@
 
 
 
+#include "../SkaleCommon.h"
 #include "boost/filesystem.hpp"
 #include "../Agent.h"
 #include "../thirdparty/json.hpp"
@@ -36,24 +37,6 @@
 
 class ConsensusBLSPublicKey;
 class ConsensusBLSPrivateKeyShare;
-
-
-/**
- * Through this interface Consensus interacts with the rest of the system
- */
-class ConsensusExtFace {
-public:
-    typedef std::vector< std::vector< uint8_t > > transactions_vector;
-
-    // Returns hashes and bytes of new transactions; blocks if there are no txns
-    virtual transactions_vector pendingTransactions( size_t _limit ) = 0;
-    // Creates new block with specified transactions AND removes them from the queue
-    virtual void createBlock(const transactions_vector &_approvedTransactions, uint64_t _timeStamp,
-            uint32_t _timeStampMillis, uint64_t _blockID,u256 _gasPrice) = 0;
-    virtual ~ConsensusExtFace() = default;
-
-    virtual void terminateApplication() {};
-};
 
 #include <boost/multiprecision/cpp_int.hpp>
 
@@ -141,5 +124,5 @@ public:
     const string &getBlsPublicKey4() const;
     const string &getBlsPrivateKey() const;
 
-    u256 getPriceForBlockId(uint64_t _blockId) override ;
+    u256 getPriceForBlockId(uint64_t _blockId) const override ;
 };
